requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-writable-nested==0.7.0
exceptiongroup==1.2.0
Faker==20.1.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
rpds-py==0.13.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
importlib-metadata